fix: fix number convert when string empty
