+++ apps/dokploy/components/ui/input.tsx
@@ -31,4 +31,37 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(
 );
 Input.displayName = "Input";
 
-export { Input };
+const NumberInput = React.forwardRef<HTMLInputElement, InputProps>(
+	({ className, errorMessage, ...props }, ref) => {
+		return (
+			<Input
+				type="text"
+				className={cn("text-left", className)}
+				ref={ref}
+				{...props}
+				onChange={(e) => {
+					const value = e.target.value;
+					if (value === "") {
+						props.onChange?.(e);
+					} else {
+						const number = Number.parseInt(value, 10);
+						if (!Number.isNaN(number)) {
+							const syntheticEvent = {
+								...e,
+								target: {
+									...e.target,
+									value: number.toString(),
+								},
+							};
+							props.onChange?.(syntheticEvent as React.ChangeEvent<HTMLInputElement>);
+						}
+					}
+				}}
+			/>
+		);
+	}
+);
+NumberInput.displayName = "NumberInput";
+
+
+export { Input, NumberInput };
